Make Atomic subclasses consistently call their parent constructors
