+++ src/Psalm/Type/Atomic/TLiteralFloat.php
@@ -15,7 +15,7 @@ final class TLiteralFloat extends TFloat
     public function __construct(float $value, bool $from_docblock = false)
     {
         $this->value = $value;
-        $this->from_docblock = $from_docblock;
+        parent::__construct($from_docblock);
     }
 
     public function getKey(bool $include_extra = true): string
