+++ src/Psalm/Type/Atomic/TLiteralInt.php
@@ -15,7 +15,7 @@ final class TLiteralInt extends TInt
     public function __construct(int $value, bool $from_docblock = false)
     {
         $this->value = $value;
-        $this->from_docblock = $from_docblock;
+        parent::__construct($from_docblock);
     }
 
     public function getKey(bool $include_extra = true): string
