+++ src/Psalm/Type/Atomic/TValueOf.php
@@ -27,7 +27,7 @@ final class TValueOf extends Atomic
     public function __construct(Union $type, bool $from_docblock = false)
     {
         $this->type = $type;
-        $this->from_docblock = $from_docblock;
+        parent::__construct($from_docblock);
     }
 
     public function getKey(bool $include_extra = true): string
